requirements.txt
selenium==4.1.0
webdriver-manager==3.5.2
pyaml==21.10.1